A Deauth packet spammer
With the correct configuration of the radiotap and and mac headers it is possible to dos a target network without being connected via spoofing deauth packets to the broadcast adress using the source mac address of the target networks wireless router. A packet injection capable wifi is required to run this program and the radiotap header bytes must be correctly set to that of the target network. THIS PROGRAM IS STRICTLY FOR EDUCATIONAL PURPOSES USE AT YOUR OWN RISK.